+++ src/save.c
@@ -1,7 +1,7 @@
 #include <save.h>
 #include <perlin.h>
 
-level *loadSave(const char *fileName, respawn **respawnList, player *player, storage *storage, int portal[4][2], int *l, item **itemList, int nItem, resource **resourceList, size_t nResource, monster **monsterList, size_t nMonster){
+level *loadSave(const char *fileName, respawn **respawnList, player *player, storage *storage, int portal[4][2], int *l, item **itemList, size_t nItem, resource **resourceList, size_t nResource, monster **monsterList, size_t nMonster){
 	char buf[256];
 	FILE *f = fopen(fileName, "r");
 	if(f == NULL){
